Time: 2 ms (40.40%) | Memory: 14.2 MB (7.44%) - LeetSync

+++ 137-single-number-ii/single-number-ii.cpp
@@ -9,6 +9,6 @@ public:
         }
         long long int thrice_sum = 3*accumulate(unique_nums_vec.begin(), unique_nums_vec.end(), 0LL);
         long long int arr_sum = accumulate(nums.begin(), nums.end(), 0LL);
-        return (thrice_sum - arr_sum)/2;
+        return (int)((thrice_sum - arr_sum)/2);
     }
 };
